[Badge] Extract component from GridItem
